move function to common
